Petals Around the Rose - Python 3
This is a first attempt at creating Petals Around the Rose in Python. There are numerous areas that will be improved on the next commit. My hope is to cut the numbers of lines down as much as possible. I bet there's a Pythonic 3 liner out there. I'm not there, yet.