Create diceroll.py
The dice roller. This currently has a global variable dicenums that ends up being accessed by the main, game logic program. Still trying to figure out how to get rid of this as I want the dice roller to be a single, callable, and reusable bit of code without knowing magical globals.